+++ src/plugin/Routes.php
@@ -55,9 +55,9 @@ trait Routes
             $event->rules['commerce/settings/producttypes/<productTypeId:\d+>'] = 'commerce/product-types/edit-product-type';
             $event->rules['commerce/settings/producttypes/new'] = 'commerce/product-types/edit-product-type';
 
-            $event->rules['commerce/settings/subscriptions/plans'] = 'commerce/plans/plan-index';
-            $event->rules['commerce/settings/subscriptions/plan/<planId:\d+>'] = 'commerce/plans/edit-plan';
-            $event->rules['commerce/settings/subscriptions/plan/new'] = 'commerce/plans/edit-plan';
+            $event->rules['commerce/settings/subscriptions/plans'] = 'commerce/plans/redirect';
+            $event->rules['commerce/settings/subscriptions/plan/<planId:\d+>'] = 'commerce/plans/redirect';
+            $event->rules['commerce/settings/subscriptions/plan/new'] = 'commerce/plans/redirect';
 
             $event->rules['commerce/orders'] = 'commerce/orders/order-index';
             $event->rules['commerce/orders/<orderId:\d+>'] = 'commerce/orders/edit-order';
@@ -129,6 +129,10 @@ trait Routes
             $event->rules['commerce/shipping/shippingmethods/<methodId:\d+>/shippingrules/new'] = 'commerce/shipping-rules/edit';
             $event->rules['commerce/shipping/shippingmethods/<methodId:\d+>/shippingrules/<ruleId:\d+>'] = 'commerce/shipping-rules/edit';
 
+            // Subscription plans
+            $event->rules['commerce/store-settings/subscription-plans'] = 'commerce/plans/plan-index';
+            $event->rules['commerce/store-settings/subscription-plans/plan/<planId:\d+>'] = 'commerce/plans/edit-plan';
+            $event->rules['commerce/store-settings/subscription-plans/plan/new'] = 'commerce/plans/edit-plan';
 
             // Taxes
             $event->rules['commerce/tax/taxcategories'] = 'commerce/tax-categories/index';